unwrapping problems with circular references
