+++ lib9ml/python/nineml/user_layer/components/interface.py
@@ -5,7 +5,6 @@ from operator import and_
 from ..base import BaseULObject, E, NINEML
 from ..utility import check_tag
 from ..random import RandomDistribution
-from . import get_or_create_component
 
 
 class Parameter(BaseULObject):
